remove route-based dialogs
