+++ app/routes/settings+/profile.photo.tsx
@@ -1,35 +1,31 @@
 import { conform, useForm } from '@conform-to/react'
 import { getFieldsetConstraint, parse } from '@conform-to/zod'
 import {
-	type DataFunctionArgs,
 	json,
 	redirect,
 	unstable_createMemoryUploadHandler,
 	unstable_parseMultipartFormData,
+	type DataFunctionArgs,
 } from '@remix-run/node'
 import {
 	Form,
-	Link,
 	useActionData,
 	useFetcher,
 	useLoaderData,
-	useNavigate,
 } from '@remix-run/react'
 import { useState } from 'react'
 import { z } from 'zod'
+import { ErrorList } from '~/components/forms.tsx'
 import { Button } from '~/components/ui/button.tsx'
-import {
-	Dialog,
-	DialogClose,
-	DialogContent,
-	DialogTitle,
-} from '~/components/ui/dialog.tsx'
+import { Icon } from '~/components/ui/icon.tsx'
 import * as deleteImageRoute from '~/routes/resources+/delete-image.tsx'
 import { authenticator, requireUserId } from '~/utils/auth.server.ts'
 import { prisma } from '~/utils/db.server.ts'
-import { ErrorList } from '~/components/forms.tsx'
-import { getUserImgSrc } from '~/utils/misc.ts'
-import { Icon } from '~/components/ui/icon.tsx'
+import { getUserImgSrc, useDoubleCheck } from '~/utils/misc.ts'
+
+export const handle = {
+	breadcrumb: <Icon name="avatar">Photo</Icon>,
+}
 
 const MAX_SIZE = 1024 * 1024 * 3 // 3MB
 
@@ -116,12 +112,14 @@ export async function action({ request }: DataFunctionArgs) {
 	return redirect('/settings/profile')
 }
 
-export default function PhotoChooserModal() {
+export default function PhotoRoute() {
 	const data = useLoaderData<typeof loader>() || {}
-	const [newImageSrc, setNewImageSrc] = useState<string | null>(null)
-	const navigate = useNavigate()
+
+	const doubleCheckDeleteImage = useDoubleCheck()
+
 	const deleteImageFetcher = useFetcher<typeof deleteImageRoute.action>()
 	const actionData = useActionData<typeof action>()
+
 	const [form, { photoFile }] = useForm({
 		id: 'profile-photo',
 		constraint: getFieldsetConstraint(PhotoFormSchema),
@@ -131,105 +129,85 @@ export default function PhotoChooserModal() {
 		},
 		shouldRevalidate: 'onBlur',
 	})
-	const [open, setOpen] = useState(true)
+
+	const [newImageSrc, setNewImageSrc] = useState<string | null>(null)
 
 	const deleteProfilePhotoFormId = 'delete-profile-photo'
-	const dismissModal = () => {
-		setOpen(false)
-		navigate('..', { preventScrollReset: true })
-	}
 	return (
-		<Dialog open={open}>
-			<DialogContent
-				onEscapeKeyDown={dismissModal}
-				onPointerDownOutside={dismissModal}
-				className="fixed left-1/2 top-1/2 w-[90vw] max-w-3xl -translate-x-1/2 -translate-y-1/2 transform rounded-lg border-2 bg-background p-12 shadow-lg"
+		<div>
+			<Form
+				method="POST"
+				encType="multipart/form-data"
+				className="flex flex-col items-center justify-center gap-10"
+				onReset={() => setNewImageSrc(null)}
+				{...form.props}
 			>
-				<DialogTitle asChild className="text-center">
-					<h2 className="text-h2">Profile photo</h2>
-				</DialogTitle>
-				<Form
-					method="POST"
-					encType="multipart/form-data"
-					className="mt-8 flex flex-col items-center justify-center gap-10"
-					onReset={() => setNewImageSrc(null)}
-					{...form.props}
-					onSubmit={dismissModal}
-				>
-					<img
-						src={
-							newImageSrc ??
-							(data.user ? getUserImgSrc(data.user?.imageId) : '')
-						}
-						className="h-64 w-64 rounded-full"
-						alt={data.user?.name ?? data.user?.username}
-					/>
-					<ErrorList errors={photoFile.errors} id={photoFile.id} />
-					<input
-						{...conform.input(photoFile, { type: 'file' })}
-						type="file"
-						accept="image/*"
-						className="sr-only"
-						tabIndex={newImageSrc ? -1 : 0}
-						onChange={e => {
-							const file = e.currentTarget.files?.[0]
-							if (file) {
-								const reader = new FileReader()
-								reader.onload = event => {
-									setNewImageSrc(event.target?.result?.toString() ?? null)
-								}
-								reader.readAsDataURL(file)
+				<img
+					src={
+						newImageSrc ?? (data.user ? getUserImgSrc(data.user?.imageId) : '')
+					}
+					className="h-52 w-52 rounded-full object-cover"
+					alt={data.user?.name ?? data.user?.username}
+				/>
+				<ErrorList errors={photoFile.errors} id={photoFile.id} />
+				<input
+					{...conform.input(photoFile, { type: 'file' })}
+					accept="image/*"
+					className="sr-only"
+					tabIndex={newImageSrc ? -1 : 0}
+					onChange={e => {
+						const file = e.currentTarget.files?.[0]
+						if (file) {
+							const reader = new FileReader()
+							reader.onload = event => {
+								setNewImageSrc(event.target?.result?.toString() ?? null)
 							}
-						}}
-					/>
-					{newImageSrc ? (
-						<div className="flex gap-4">
-							<Button type="submit">Save Photo</Button>
-							<Button type="reset" variant="secondary">
-								Reset
+							reader.readAsDataURL(file)
+						}
+					}}
+				/>
+				{newImageSrc ? (
+					<div className="flex gap-4">
+						<Button type="submit">Save Photo</Button>
+						<Button type="reset" variant="secondary">
+							Reset
+						</Button>
+					</div>
+				) : (
+					<div className="flex gap-4">
+						<Button asChild className="cursor-pointer">
+							<label htmlFor={photoFile.id}>
+								<Icon name="pencil-1">Change</Icon>
+							</label>
+						</Button>
+						{data.user?.imageId ? (
+							<Button
+								variant="destructive"
+								{...doubleCheckDeleteImage.getButtonProps({
+									type: 'submit',
+									form: deleteProfilePhotoFormId,
+								})}
+							>
+								<Icon name="trash">
+									{doubleCheckDeleteImage.doubleCheck
+										? 'Are you sure?'
+										: 'Delete'}
+								</Icon>
 							</Button>
-						</div>
-					) : (
-						<div className="flex gap-4">
-							<Button asChild className="cursor-pointer">
-								<label htmlFor={photoFile.id} className="flex gap-1">
-									<Icon name="pencil-1" /> Change
-								</label>
-							</Button>
-							{data.user?.imageId ? (
-								<Button
-									variant="destructive"
-									type="submit"
-									form={deleteProfilePhotoFormId}
-									className="flex gap-1"
-								>
-									<Icon name="trash" /> Delete
-								</Button>
-							) : null}
-						</div>
-					)}
-					<ErrorList errors={form.errors} />
-				</Form>
-				<DialogClose asChild>
-					<Link
-						to=".."
-						onClick={() => setOpen(false)}
-						preventScrollReset
-						aria-label="Close"
-						className="absolute right-10 top-10"
-					>
-						<Icon name="cross-1" />
-					</Link>
-				</DialogClose>
-			</DialogContent>
+						) : null}
+					</div>
+				)}
+				<ErrorList errors={form.errors} />
+			</Form>
 			<deleteImageFetcher.Form
 				method="POST"
 				id={deleteProfilePhotoFormId}
 				action={deleteImageRoute.ROUTE_PATH}
+				onSubmit={() => setNewImageSrc(null)}
 			>
 				<input name="intent" type="hidden" value="submit" />
 				<input name="imageId" type="hidden" value={data.user?.imageId ?? ''} />
 			</deleteImageFetcher.Form>
-		</Dialog>
+		</div>
 	)
 }
