+++ package.json
@@ -41,7 +41,6 @@
     "@conform-to/zod": "^0.7.2",
     "@prisma/client": "^5.0.0",
     "@radix-ui/react-checkbox": "^1.0.4",
-    "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-slot": "^1.0.2",
